refactor(vscode): isolate project config prompts
